Add files via upload
My networking diagram includes the internet, a wireless router, Pc, laptop, and printer. This is an example of a star topology. This is probably the most common in households now due to so many people having multiple wireless devices. With the help of a router directing the traffic, multiple devices can be wirelessly connected to it at the same time.